refactor to metadata extractor for reuse
